refactor: move tags query from package to client workspace
