+++ client/src/data-provider/index.ts
@@ -3,3 +3,4 @@ export * from './mutations';
 export * from './prompts';
 export * from './queries';
 export * from './roles';
+export * from './tags';
