+++ client/src/data-provider/tags.ts
@@ -1,0 +1,19 @@
+import { useQuery } from '@tanstack/react-query';
+import type { UseQueryOptions, QueryObserverResult } from '@tanstack/react-query';
+import type { TConversationTagsResponse } from 'librechat-data-provider';
+import { QueryKeys, dataService } from 'librechat-data-provider';
+
+export const useGetConversationTags = (
+  config?: UseQueryOptions<TConversationTagsResponse>,
+): QueryObserverResult<TConversationTagsResponse> => {
+  return useQuery<TConversationTagsResponse>(
+    [QueryKeys.conversationTags],
+    () => dataService.getConversationTags(),
+    {
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+      refetchOnMount: false,
+      ...config,
+    },
+  );
+};
